Add operation, calculator service

+++ src/main/java/com/example2/controller/OnlineCalculatorController.java
@@ -25,7 +25,7 @@ public class OnlineCalculatorController {
 		Operation[] operations = arithmaticParser.parse(input);
 
 		// calculate result
-		int result = calculatorService.calculate(operations);
+		double result = calculatorService.calculate(operations);
 
 		// format response
 		String response = responseFormatter.format(result);
